Animate new roll results
